+++ x/evidence/client/cli/query_test.go
@@ -10,9 +10,9 @@ import (
 
 	"cosmossdk.io/x/evidence"
 	"cosmossdk.io/x/evidence/client/cli"
+	abci "github.com/cometbft/cometbft/abci/types"
+	rpcclientmock "github.com/cometbft/cometbft/rpc/client/mock"
 	"github.com/stretchr/testify/require"
-	abci "github.com/tendermint/tendermint/abci/types"
-	rpcclientmock "github.com/tendermint/tendermint/rpc/client/mock"
 
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"