+++ x/staking/module_test.go
@@ -3,8 +3,8 @@ package staking_test
 import (
 	"testing"
 
+	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 	"github.com/stretchr/testify/require"
-	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
 	authKeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"